+++ metadata-ingestion/src/datahub/sql_parsing/schema_resolver.py
@@ -1,0 +1,191 @@
+import contextlib
+import pathlib
+from typing import Dict, List, Optional, Set, Tuple
+
+from typing_extensions import TypedDict
+
+from datahub.emitter.mce_builder import (
+    DEFAULT_ENV,
+    make_dataset_urn_with_platform_instance,
+)
+from datahub.ingestion.api.closeable import Closeable
+from datahub.ingestion.graph.client import DataHubGraph
+from datahub.ingestion.source.bigquery_v2.bigquery_audit import BigqueryTableIdentifier
+from datahub.metadata.schema_classes import SchemaFieldClass, SchemaMetadataClass
+from datahub.sql_parsing._models import _TableName
+from datahub.sql_parsing.sql_parsing_common import TABLE_CASE_SENSITIVE_PLATFORMS
+from datahub.utilities.file_backed_collections import ConnectionWrapper, FileBackedDict
+from datahub.utilities.urns.field_paths import get_simple_field_path_from_v2_field_path
+
+# A lightweight table schema: column -> type mapping.
+SchemaInfo = Dict[str, str]
+
+
+class GraphQLSchemaField(TypedDict):
+    fieldPath: str
+    nativeDataType: str
+
+
+class GraphQLSchemaMetadata(TypedDict):
+    fields: List[GraphQLSchemaField]
+
+
+class SchemaResolver(Closeable):
+    def __init__(
+        self,
+        *,
+        platform: str,
+        platform_instance: Optional[str] = None,
+        env: str = DEFAULT_ENV,
+        graph: Optional[DataHubGraph] = None,
+        _cache_filename: Optional[pathlib.Path] = None,
+    ):
+        # TODO handle platforms when prefixed with urn:li:dataPlatform:
+        self.platform = platform
+        self.platform_instance = platform_instance
+        self.env = env
+
+        self.graph = graph
+
+        # Init cache, potentially restoring from a previous run.
+        shared_conn = None
+        if _cache_filename:
+            shared_conn = ConnectionWrapper(filename=_cache_filename)
+        self._schema_cache: FileBackedDict[Optional[SchemaInfo]] = FileBackedDict(
+            shared_connection=shared_conn,
+        )
+
+    def get_urns(self) -> Set[str]:
+        return set(self._schema_cache.keys())
+
+    def schema_count(self) -> int:
+        return len(self._schema_cache)
+
+    def get_urn_for_table(self, table: _TableName, lower: bool = False) -> str:
+        # TODO: Validate that this is the correct 2/3 layer hierarchy for the platform.
+
+        table_name = ".".join(
+            filter(None, [table.database, table.db_schema, table.table])
+        )
+
+        platform_instance = self.platform_instance
+
+        if lower:
+            table_name = table_name.lower()
+            platform_instance = platform_instance.lower() if platform_instance else None
+
+        if self.platform == "bigquery":
+            # Normalize shard numbers and other BigQuery weirdness.
+            with contextlib.suppress(IndexError):
+                table_name = BigqueryTableIdentifier.from_string_name(
+                    table_name
+                ).get_table_name()
+
+        urn = make_dataset_urn_with_platform_instance(
+            platform=self.platform,
+            platform_instance=platform_instance,
+            env=self.env,
+            name=table_name,
+        )
+        return urn
+
+    def resolve_table(self, table: _TableName) -> Tuple[str, Optional[SchemaInfo]]:
+        urn = self.get_urn_for_table(table)
+
+        schema_info = self._resolve_schema_info(urn)
+        if schema_info:
+            return urn, schema_info
+
+        urn_lower = self.get_urn_for_table(table, lower=True)
+        if urn_lower != urn:
+            schema_info = self._resolve_schema_info(urn_lower)
+            if schema_info:
+                return urn_lower, schema_info
+
+        if self.platform in TABLE_CASE_SENSITIVE_PLATFORMS:
+            return urn, None
+        else:
+            return urn_lower, None
+
+    def has_urn(self, urn: str) -> bool:
+        return urn in self._schema_cache
+
+    def _resolve_schema_info(self, urn: str) -> Optional[SchemaInfo]:
+        if urn in self._schema_cache:
+            return self._schema_cache[urn]
+
+        # TODO: For bigquery partitioned tables, add the pseudo-column _PARTITIONTIME
+        # or _PARTITIONDATE where appropriate.
+
+        if self.graph:
+            schema_info = self._fetch_schema_info(self.graph, urn)
+            if schema_info:
+                self._save_to_cache(urn, schema_info)
+                return schema_info
+
+        self._save_to_cache(urn, None)
+        return None
+
+    def add_schema_metadata(
+        self, urn: str, schema_metadata: SchemaMetadataClass
+    ) -> None:
+        schema_info = self._convert_schema_aspect_to_info(schema_metadata)
+        self._save_to_cache(urn, schema_info)
+
+    def add_raw_schema_info(self, urn: str, schema_info: SchemaInfo) -> None:
+        self._save_to_cache(urn, schema_info)
+
+    def add_graphql_schema_metadata(
+        self, urn: str, schema_metadata: GraphQLSchemaMetadata
+    ) -> None:
+        schema_info = self.convert_graphql_schema_metadata_to_info(schema_metadata)
+        self._save_to_cache(urn, schema_info)
+
+    def _save_to_cache(self, urn: str, schema_info: Optional[SchemaInfo]) -> None:
+        self._schema_cache[urn] = schema_info
+
+    def _fetch_schema_info(self, graph: DataHubGraph, urn: str) -> Optional[SchemaInfo]:
+        aspect = graph.get_aspect(urn, SchemaMetadataClass)
+        if not aspect:
+            return None
+
+        return self._convert_schema_aspect_to_info(aspect)
+
+    @classmethod
+    def _convert_schema_aspect_to_info(
+        cls, schema_metadata: SchemaMetadataClass
+    ) -> SchemaInfo:
+        return cls._convert_schema_field_list_to_info(schema_metadata.fields)
+
+    @classmethod
+    def _convert_schema_field_list_to_info(
+        cls, schema_fields: List[SchemaFieldClass]
+    ) -> SchemaInfo:
+        return {
+            get_simple_field_path_from_v2_field_path(col.fieldPath): (
+                # The actual types are more of a "nice to have".
+                col.nativeDataType
+                or "str"
+            )
+            for col in schema_fields
+            # TODO: We can't generate lineage to columns nested within structs yet.
+            if "." not in get_simple_field_path_from_v2_field_path(col.fieldPath)
+        }
+
+    @classmethod
+    def convert_graphql_schema_metadata_to_info(
+        cls, schema: GraphQLSchemaMetadata
+    ) -> SchemaInfo:
+        return {
+            get_simple_field_path_from_v2_field_path(field["fieldPath"]): (
+                # The actual types are more of a "nice to have".
+                field["nativeDataType"]
+                or "str"
+            )
+            for field in schema["fields"]
+            # TODO: We can't generate lineage to columns nested within structs yet.
+            if "." not in get_simple_field_path_from_v2_field_path(field["fieldPath"])
+        }
+
+    def close(self) -> None:
+        self._schema_cache.close()
