+++ metadata-ingestion/tests/performance/bigquery/test_bigquery_usage.py
@@ -13,7 +13,7 @@ from datahub.ingestion.source.bigquery_v2.bigquery_config import (
 )
 from datahub.ingestion.source.bigquery_v2.bigquery_report import BigQueryV2Report
 from datahub.ingestion.source.bigquery_v2.usage import BigQueryUsageExtractor
-from datahub.sql_parsing.sqlglot_lineage import SchemaResolver
+from datahub.sql_parsing.schema_resolver import SchemaResolver
 from datahub.utilities.perf_timer import PerfTimer
 from tests.performance.bigquery.bigquery_events import generate_events, ref_from_table
 from tests.performance.data_generation import (
