+++ metadata-ingestion/tests/unit/sql_parsing/goldens/test_select_max_with_schema.json
@@ -1,7 +1,7 @@
 {
     "query_type": "SELECT",
     "query_type_props": {},
-    "query_fingerprint": "ca9df3c737a7275934a0d3950bc76169f36a5931ccc05967b394849ea1d1863f",
+    "query_fingerprint": "727911301a80a595a80671e784793aa2ea1be2e2924e10eba3e0c622e5082b0d",
     "in_tables": [
         "urn:li:dataset:(urn:li:dataPlatform:mysql,mytable,PROD)"
     ],
