+++ metadata-ingestion/tests/unit/sql_parsing/test_schemaresolver.py
@@ -1,4 +1,4 @@
-from datahub.sql_parsing.sqlglot_lineage import SchemaResolver, _TableName
+from datahub.sql_parsing.schema_resolver import SchemaResolver, _TableName
 
 
 def test_get_urn_for_table_lowercase():
