+++ metadata-ingestion/tests/unit/sql_parsing/test_sqlglot_lineage.py
@@ -33,8 +33,7 @@ FROM mytable
                 "col2": "NUMBER",
             },
         },
-        # Shared with the test above.
-        expected_file=RESOURCE_DIR / "test_select_max.json",
+        expected_file=RESOURCE_DIR / "test_select_max_with_schema.json",
     )
