+++ metadata-ingestion/tests/unit/test_bigquery_lineage.py
@@ -14,7 +14,7 @@ from datahub.ingestion.source.bigquery_v2.lineage import (
     BigqueryLineageExtractor,
     LineageEdge,
 )
-from datahub.sql_parsing.sqlglot_lineage import SchemaResolver
+from datahub.sql_parsing.schema_resolver import SchemaResolver
 
 
 @pytest.fixture
