+++ metadata-ingestion/tests/unit/test_bigquery_sql_lineage.py
@@ -1,7 +1,8 @@
 from typing import List
 
 from datahub.ingestion.source.bigquery_v2.bigquery_audit import BigQueryTableRef
-from datahub.sql_parsing.sqlglot_lineage import SchemaResolver, sqlglot_lineage
+from datahub.sql_parsing.schema_resolver import SchemaResolver
+from datahub.sql_parsing.sqlglot_lineage import sqlglot_lineage
 
 
 class BigQuerySQLParser:
